Resolves the nameing of sections and marginally improved teh UI on creation.
Nexuses now use the GUID, but this does not leak into the UI.
The link tool as deployed by default uses a default group/section which is the title of the 
worksite, however if link tool has a config property "exegesis-section" set that will
be the section used for that invocation of the tool. The section should be of the for
sitetitle.grouptitle

If you lainch the tool you can see this list from the dropdown on the Discuss page.