Alison Pearn made amendments including those below on 4 Feb 2008. These need merging with the trunk asap.
v01_letters.xml

Calendar no. 209, to Herbert, J. M. 
Some of the tex markup has failed to convert, including the code for the calendar number itself.  Fix by hand.  See vol. 2 pp. 317-19 to check text.
NB - the single footnote marker needs fixing from 208.f1 to 209.f1

This problem occurred because the previous letter had no footnotes and the end of the letter wasn't apparent to the conversion programme.   At the moment the text of this one is appearing at the bottom of letter 208.  The footnotes for 208 are not properly delimited.


Asa_Gray_letters.xml

Cal no 13028 is similarly incorporated at the bottom of 12921