Removed messages between developers (probably from Andy Thornton to Ian Boston or Alison Pearn) from the code so that they would no longer be public facing(!). Removed some duplicate or otherwise not used/useful described queries.
Note: described-queries.xml is not well-formed. It wasn't before I edited it, and I have left that aspect of it unchanged rather than risk breaking the next Continuum build. Ideally, any code that parses described-queries.xml should be checked for its ability to handle well-formed XML. If all such code CAN handle well-formed XML, then described-queries.xml should be amended to be well-formed (specifically, < and > need to be replaced with &lt; and &gt; ).