Added latest proofread belief letters to belief_letts.xml .
NB. There are undeclared entities referenced in these letters. This is due to the DCP's parse-to-xml.pl script being out-of-sync with darmll.dtd (and darmll.dtd's subsidiary files). (To compound matters, the latest copy of darmll.dtd (and subsidiaries) at the DCP may also be out-of-sync with the latest copy at CARET.) I have cautioned against running a Continuum build until (i) all entities referenced in the file have been declared in CARET's copy of darmll.dtd (and its subsidiaries) and (ii) either the DCP's copy has been declared deprecated, or the DCP's copy has been updated to mirror CARET's.