OK: log numbers:  I've found them all I think. Would you mind making the changes as I can't get at the files at the moment? Thanks.
This first one is an outright mistake -

log num 8322 occurs twice in calendar_entries_12000-12999.xml
> (calendarnums
cal 12284 = log 7264
cal 12285 = log 8322


The rest are the result of cancelling calendar numbers - this happens when we realise that one "letter" is actually part of another, or when we've just failed to realise we already have it.  I suppose from a databasy point of view it would be a good idea to put in the log number of the "new" calendar number.  If I was cancelling an entry now, I wouldn't delete the original log number, but that has happened with these and so long ago it is lost in the mists...


> log num 19389 occurs twice in calendar_entries_04000-04999.xml
> (calendarnums
cal 4822 = 19389
cal 4821 = cancelled:  part of cal 6332a log  20772
>

> log num [!?missing] occurs 5 times (calendarnums
cal 1049 = (cancelled: same as cal 1083 log 3132)
cal 10388f = (cancelled: same as cal 10840 log 19091)
cal 569 = (cancelled: same as cal 555 log 1850)
cal 645 = (cancelled: same as cal 649 log 28)
cal 936 = (cancelled: same as cal 1014 log 1852)