Completely refactored & cleaned up text monster. Lots of bugs fixed and features implemented.
All the code bas been re-organised to be easier to manage in the future (I 
hope). I've removed lots of old (now dead) code. It's still (by default) 
configured to run on my solr schema (the data generated by darwin2solr.xslt in 
the solr-data dir), so the next job is to modify things to work with Chris's
solr data. Should be a case of specifying the field names in a new config file.