Added error handling pages which are shown on 500 and 404 HTTP responses.
The client javascript now has a .error() handler on the main search ajax call.
The server has an error handler for 500 and 404 errors which shows an error 
message. In keeping with the consistency between normal and ajax requests, both
will display identical errors to the client, regardless of whether the search
was initiated by an ajax request or a normal browser request.

In doing this I also discovered another error: the provenance field was being
parsed as XML, but it should have been treated as plain text. As a result 
occasionally the parse would fail, for example if an ampersand was present in 
the text.