Added support for serving Bitstreams from the upstream proxy server using the X-sendfile header mechanism supported by Apache Httpd, nginx, litehttpd and others.
When this is used the BitStreamRedirectReader will create a signed URL that redirects to the target http virtual host. That request proxes back to the appropriate XSendfile*Servlet which  if the request is valid, responds with headers telling the reverse proxy server where the file is. Once that is done, the reverse proxy server serves the file directly from the filesystem and the backend server is free to continue with other tasks. This technique can be used with any server that supports the mechanism and can be used to direct to WebSocket type end points with security.