Rewrite cudl-proxy.php to use CUDL metadata JSON
It fetches the CUDL metadata for an item and grabs the .dzi URL from
that instead of assembling the item+page ID and .dzi URL ourselves.

This should make us more resilient to changes on the CUDL end.

Note that cURL support for PHP is not installed on our server, so the
code is currently falling back to the old proxy until that's resolved.