Update RemoteUserLoginView to not rely on RemoteUserMiddleware.
It turns out that RemoteUserMiddleware expects the entire site to be
protected by the webserver, so if any page doesn't have a REMOTE_USER
var set the user gets logged out.... Which is why the user was only
logged in on the login page itself and nowhere else. :)

RemoteUserLoginView basically replaces RemoteUserMiddleware for the use
case where you only have the webserver protect the login url itself, not
other pages.