Tweak print css to avoid calendar getting cut off
We were relying on javascript to resize the calendar before printing.

Chrome has a bug causing such before-print hooks to fail:
https://code.google.com/p/chromium/issues/detail?id=401179

Firefox was cutting off the calendar, even though it was getting
resized.

Instead of all this, we'll just resize the calendar in the print css
to be 100% of the page width. This seems to work well with
FullCalendar 2.x (quite possibly wouldn't have done in the past with
1.x).